+++ gjtjb/RKWXT/VirtualMall/VirtualGoodsInfo/Model/VirtualOrderInfoEntity.h
@@ -15,7 +15,7 @@
 @property (nonatomic,assign)CGFloat postage;
 @property (nonatomic,assign)CGFloat backMoney;
 @property (nonatomic,assign)NSInteger stockID;
-@property (nonatomic,assign)CGFloat xnbPrice;
+@property (nonatomic,assign)int xnbPrice;
 @property (nonatomic,strong) NSString *goodsImg;
 @property (nonatomic,strong) NSString *stockName;
 @property (nonatomic,strong) NSString *goodsName;
